requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
bleach==6.1.0
Django==5.1.2
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
django-summernote==0.8.20.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-yasg==1.21.8
et-xmlfile==1.1.0
hiredis==3.0.0
inflection==0.5.1
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
webencodings==0.5.1
